feat: improve testability by moving FeatureResult to public module
