+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "growthbook-rust-sdk"
-version = "1.0.0"
+version = "1.1.0"
 edition = "2021"
 authors = [
     "Deroldo <diogo.fernandes@willbank.com.br",
@@ -26,8 +26,8 @@ reqwest = { version = "0.12.4", default-features = false, features = [
 ] }
 reqwest-middleware = { version = "0.4.0" }
 
-strum = { version = "0.26.1" }
-strum_macros = { version = "0.26.1" }
+strum = { version = "0.27.1" }
+strum_macros = { version = "0.27.1" }
 
 serde = { version = "1.0.196", features = ["derive"] }
 serde_json = { version = "1.0.113", features = ["preserve_order"] }
@@ -41,12 +41,12 @@ regex = { version = "1.10.4" }
 indexmap = { version = "2.2.6" }
 
 [dev-dependencies]
-tower = { version = "0.4.13", features = ["util"] }
+tower = { version = "0.5.2", features = ["util"] }
 
-rstest = { version = "0.21.0" }
-test-context = { version = "0.3.0" }
+rstest = { version = "0.25.0" }
+test-context = { version = "0.4.1" }
 
 wiremock = { version = "0.6.0" }
 
-rand = { version = "0.8.5" }
+rand = { version = "0.9.1" }
 uuid = { version = "1.8.0", features = ["v7", "serde"] }
