+++ examples/client/src/main.rs
@@ -1,7 +1,9 @@
 use growthbook_rust_sdk::client::GrowthBookClient;
+use growthbook_rust_sdk::model_public::{FeatureResult, GrowthBookAttribute};
+use serde::Deserialize;
+use serde_json::json;
 use std::time::Duration;
 use tokio::time::sleep;
-use serde::Deserialize;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -22,13 +24,17 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
             let string_feature = gb_client.feature_result(feature_name, None);
             let string = string_feature.value_as::<String>()?;
-            println!("feature: {} string value {:?}", feature_name, string)
+            println!("feature: {} string value {:?}", feature_name, string);
 
             let custom_feature = gb_client.feature_result(feature_name, None);
             let custom = custom_feature.value_as::<Custom>()?;
             println!("feature: {} custom first value {:?}", feature_name, custom.first);
             println!("feature: {} custom second value {:?}", feature_name, custom.second);
         }
+
+        // Example of creating FeatureResult for testing/mocking
+        let test_result = FeatureResult::new(json!("test-value"), true, "test".to_string());
+        println!("Test FeatureResult: on={}, value={}", test_result.on, test_result.value);
 
         sleep(Duration::from_secs(5)).await;
     }
