+++ src/lib.rs
@@ -11,7 +11,6 @@ mod gateway;
 mod growthbook;
 mod hash;
 mod infra;
-mod model_private;
 pub mod model_public;
 pub mod namespace;
 mod range;
